moving sneak & sprint properties to Living
