+++ tests/phpstan/configs/l8-baseline.neon
@@ -171,16 +171,6 @@ parameters:
 			path: ../../../src/entity/Entity.php
 
 		-
-			message: "#^Cannot call method getValue\\(\\) on pocketmine\\\\entity\\\\Attribute\\|null\\.$#"
-			count: 2
-			path: ../../../src/entity/Entity.php
-
-		-
-			message: "#^Cannot call method setValue\\(\\) on pocketmine\\\\entity\\\\Attribute\\|null\\.$#"
-			count: 1
-			path: ../../../src/entity/Entity.php
-
-		-
 			message: "#^Property pocketmine\\\\network\\\\mcpe\\\\protocol\\\\AddActorPacket\\:\\:\\$type \\(string\\) does not accept string\\|null\\.$#"
 			count: 1
 			path: ../../../src/entity/Entity.php
@@ -194,6 +184,16 @@ parameters:
 			message: "#^Method pocketmine\\\\entity\\\\HungerManager\\:\\:fetchAttribute\\(\\) should return pocketmine\\\\entity\\\\Attribute but returns pocketmine\\\\entity\\\\Attribute\\|null\\.$#"
 			count: 1
 			path: ../../../src/entity/HungerManager.php
+
+		-
+			message: "#^Cannot call method getValue\\(\\) on pocketmine\\\\entity\\\\Attribute\\|null\\.$#"
+			count: 2
+			path: ../../../src/entity/Living.php
+
+		-
+			message: "#^Cannot call method setValue\\(\\) on pocketmine\\\\entity\\\\Attribute\\|null\\.$#"
+			count: 1
+			path: ../../../src/entity/Living.php
 
 		-
 			message: "#^Cannot call method getEffectLevel\\(\\) on pocketmine\\\\entity\\\\effect\\\\EffectInstance\\|null\\.$#"
